@@ -32,24 +32,24 @@
   "typings": "index.d.ts",
   "homepage": "https://github.com/Greentube/localize-router",
   "peerDependencies": {
-    "@angular/core": ">=4.0.0",
-    "@angular/common": ">=4.0.0",
-    "@angular/router": ">=4.0.0",
-    "rxjs": "^5.0.0",
-    "@ngx-translate/core": ">=8.0.0"
+    "@angular/core": ">=6.0.0",
+    "@angular/common": ">=6.0.0",
+    "@angular/router": ">=6.0.0",
+    "rxjs": "^6.0.0",
+    "@ngx-translate/core": ">=10.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "~4.3.3",
-    "@angular/common": "~4.3.3",
-    "@angular/compiler": "~4.3.3",
-    "@angular/compiler-cli": "~4.3.3",
-    "@angular/core": "~4.3.3",
-    "@angular/http": "~4.3.3",
-    "@angular/platform-browser": "~4.3.3",
-    "@angular/platform-browser-dynamic": "~4.3.3",
-    "@angular/platform-server": "~4.3.3",
-    "@angular/router": "~4.3.3",
-    "@ngx-translate/core": "8.0.0",
+    "@angular/animations": "~6.0.6",
+    "@angular/common": "~6.0.6",
+    "@angular/compiler": "~6.0.6",
+    "@angular/compiler-cli": "~6.0.6",
+    "@angular/core": "~6.0.6",
+    "@angular/http": "~6.0.6",
+    "@angular/platform-browser": "~6.0.6",
+    "@angular/platform-browser-dynamic": "~6.0.6",
+    "@angular/platform-server": "~6.0.6",
+    "@angular/router": "~6.0.6",
+    "@ngx-translate/core": "10.0.2",
     "@types/es6-promise": "0.0.32",
     "@types/hammerjs": "2.0.34",
     "@types/jasmine": "2.5.47",
@@ -70,13 +70,14 @@
     "karma-webpack": "2.0.3",
     "loader-utils": "1.1.0",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.0",
+    "rxjs": "6.2.1",
+    "rxjs-compat": "^6.2.1",
     "source-map-loader": "0.2.1",
     "ts-helpers": "1.1.2",
     "tslint": "5.8.0",
     "tslint-eslint-rules": "^4.1.1",
     "tslint-loader": "3.5.3",
-    "typescript": "2.3.3",
+    "typescript": "2.7.2",
     "webpack": "2.6.0",
     "webpack-node-externals": "^1.5.4",
     "zone.js": "0.8.11"
